修正串口读取返回值类型，将int32_t改为int64_t以支持更大数据范围

+++ private_src/serial/Serial1.cpp
@@ -323,7 +323,7 @@ bsp::Serial1::~Serial1()
 	Close();
 }
 
-int32_t bsp::Serial1::Read(base::Span const &span)
+int64_t bsp::Serial1::Read(base::Span const &span)
 {
 	if (_closed)
 	{
